Move models into a proper `model` package
